+++ call-center/web-client/src/components/Common/index.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useRef, useState, useCallback } from 'react';
-import { TelnyxRTC } from '@telnyx/webrtc';
-import { updateAgent } from '../../services/agentsService';
+
 import * as callsService from '../../services/callsService';
 import useAgents from '../../hooks/useAgents';
 import ActiveCall from '../ActiveCall';
@@ -8,113 +7,24 @@ import Agents from '../Agents';
 import Dialer from '../Dialer';
 import LoadingIcon from '../LoadingIcon';
 import ICommon from '../../interfaces/ICommon';
-import IPartialWebRTCCall from '../../interfaces/IPartialWebRTCCall';
+import useTelnyxRTC from '../../hooks/useTelnyxRTC';
 
 
-function Common({ agentId, agentSipUsername, agentName, token }: ICommon) {
-  // Save the Telnyx WebRTC client as a ref as to persist
-  // the client object through component updates
-  let telnyxClientRef = useRef<TelnyxRTC>();
+function Common({ agentId, agentSipUsername, token }: ICommon) {
+  
   let audioRef = useRef<HTMLAudioElement>(null);
-  // Check if component is mounted before updating state
-  // in the Telnyx WebRTC client callbacks
-  let isMountedRef = useRef<boolean>(false);
+ 
   let [dialingDestination, setDialingDestination] = useState<string | null>();
-  let [webRTCClientState, setWebRTCClientState] = useState<string>('');
-  let [webRTCall, setWebRTCCall] = useState<IPartialWebRTCCall | null>();
+
+  let {telnyxClientRef, webRTCall, webRTCClientState} = useTelnyxRTC(agentId, token, setDialingDestination);
+  
   let agentsState = useAgents(agentSipUsername);
 
-  const updateWebRTCState = (state: string) => {
-    if (isMountedRef.current) {
-      setWebRTCClientState(state);
-    }
-  };
+  console.log('telnyxClientRef',telnyxClientRef)
+  console.log('webRTCall',webRTCall)
+  console.log('webRTCClientState',webRTCClientState)
 
-  useEffect(() => {
-    isMountedRef.current = true;
 
-    const telnyxClient = new TelnyxRTC({
-      login_token: token,
-    });
-
-    telnyxClient.on('telnyx.ready', () => {
-      console.log('ready');
-
-      updateWebRTCState('ready');
-
-      updateAgent(agentId, { available: true });
-    });
-
-    telnyxClient.on('telnyx.error', (error: any) => {
-      console.error('error:', error);
-
-      updateWebRTCState('error');
-
-      updateAgent(agentId, { available: false });
-    });
-
-    telnyxClient.on('telnyx.socket.close', () => {
-      console.log('close');
-
-      telnyxClient.disconnect();
-
-      updateWebRTCState('disconnected');
-
-      updateAgent(agentId, { available: false });
-    });
-
-    telnyxClient.on('telnyx.notification', (notification: any) => {
-      console.log('notification:', notification);
-
-      if (notification.call) {
-        const {
-          state,
-          direction,
-          options,
-          answer,
-          hangup,
-          muteAudio,
-          unmuteAudio,
-          remoteStream,
-        } = notification.call;
-
-        console.log('state:', state);
-
-        if (state === 'hangup' || state === 'destroy') {
-          setWebRTCCall(null);
-
-          updateAgent(agentId, { available: true });
-        } else {
-          if (state === 'answering') {
-            updateAgent(agentId, { available: false });
-          } else if (state === 'active') {
-            setDialingDestination(null);
-          }
-
-          setWebRTCCall({
-            state,
-            direction,
-            options,
-            remoteStream,
-            answer: answer.bind(notification.call),
-            hangup: hangup.bind(notification.call),
-            muteAudio: muteAudio.bind(notification.call),
-            unmuteAudio: unmuteAudio.bind(notification.call),
-          });
-        }
-      }
-    });
-
-    telnyxClientRef.current = telnyxClient;
-    telnyxClientRef.current.connect();
-
-    return () => {
-      isMountedRef.current = false;
-
-      telnyxClientRef.current?.disconnect();
-      telnyxClientRef.current = undefined;
-    };
-  }, [token, agentId]);
 
   const dial = useCallback(
     (destination) => {
